user suthentificate or admin

+++ events/views.py
@@ -3,9 +3,11 @@ import calendar
 from calendar import HTMLCalendar
 from datetime import datetime
 from .models import Venue, Event
-from .forms import VenueForm, EventForm
+from .forms import VenueForm, EventForm, EventFormAdmin
+from django.contrib.auth.models import User
 from django.http import HttpResponseRedirect
 from django.http import HttpResponse
+from django.contrib import messages
 import csv
 
 # pdf things to import 
@@ -123,7 +125,10 @@ def add_venue(request):
     if request.method == "POST":
         form = VenueForm(request.POST)
         if form.is_valid():
-            form.save()
+            # form.save()
+            venue = form.save(commit=False)
+            venue.owner = request.user.id #loggrd in user
+            venue.save()
             return HttpResponseRedirect('/add_venue?submitted=True')
     else:
         form = VenueForm
@@ -146,6 +151,7 @@ def list_venues(request):
 
 def show_venue(request, venue_id):
     venue = Venue.objects.get(pk=venue_id)
+    venue_owner= User.objects.get(pk=venue.owner)
     return render(request, 'events/show_venue.html', {'venue': venue})
 
 
@@ -170,19 +176,34 @@ def update_venue(request, venue_id):
 def add_event(request):
     submitted = False
     if request.method == "POST":
-        form = EventForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect('/add_event?submitted=True')
+        if request.user.is_superuser:
+            form = EventFormAdmin(request.POST)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect('/add_event?submitted=True')
+        else:
+            form = EventForm(request.POST)
+            if form.is_valid():
+                event = form.save(commit=False)
+                event.manager = request.user
+                event.save()
+                return HttpResponseRedirect('/add_event?submitted=True')
     else:
-        form = EventForm
+        if request.user.is_superuser:
+            form = EventFormAdmin
+        else:
+            form= EventForm            
         if 'submitted' in request.GET:
             submitted = True
     return render(request, 'events/add_event.html', {'form': form, 'submitted': submitted})
+            
 
 def update_event(request, event_id):
     event = Event.objects.get(pk= event_id)
-    form = EventForm(request.POST or None, instance=event)
+    if request.user.is_superuser:
+        form = EventFormAdmin(request.POST or None, instance=event)
+    else:
+        form = EventForm(request.POST or None, instance=event)
     if form.is_valid():
         form.save()
         return redirect('list-events')
@@ -191,8 +212,14 @@ def update_event(request, event_id):
     
 def delete_event(request, event_id):
     event = Event.objects.get(pk=event_id)
-    event.delete()
-    return redirect('list-events')
+    if request.user == event.manager:
+        event.delete()
+        messages.success(request, 'You have succesfully deleted the event')
+        return redirect('list-events')
+    else: 
+        messages.error(request, 'You are Not the event manager You can Not delete the event')
+        return redirect('list-events')
+        
 
 def delete_venue(request, venue_id):
     event = Venue.objects.get(pk=venue_id)
